+++ includes/Admin/Settings.php
@@ -14,7 +14,7 @@
 
 namespace SilverAssist\ACFCloneFields\Admin;
 
-use SilverAssist\ACFCloneFields\Core\Interfaces\LoadableInterface;
+use SilverAssist\PluginKernel\Interfaces\LoadableInterface;
 use SilverAssist\SettingsHub\SettingsHub;
 
 defined( 'ABSPATH' ) || exit;